feat[api]: delete article of user (#19)
* change func IdNum's return of valueobject Id, uint64 to uint

* fix owner index error in ArticleDO, replace field Owner's tag `uniqueIndex` to `index`"

* feat[api]: delete article of user

+++ article/app/dto.go
@@ -3,7 +3,7 @@ package app
 import "victorzhou123/vicblog/article/domain"
 
 type ArticleListDto struct {
-	Id        uint64 `json:"id"`
+	Id        uint   `json:"id"`
 	Title     string `json:"title"`
 	Cover     string `json:"cover"`
 	IsPublish bool   `json:"isPublish"`
